requirements.txt
pandas
numpy
pdfplumber
python-dotenv
faiss-cpu
logging
langchain
langchain_openai
langchain_community
langchain-core
langgraph
tabulate
pyppeteer
openpyxl

